requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-filter==23.2
psycopg2-binary==2.9.5
djangorestframework-simplejwt